Add separate folders for parent and volunteer pages
